fix(message): Fixes BOT-13

+++ src/commands/moderation/purge/clean.ts
@@ -110,8 +110,10 @@ export default class extends Command {
 		}
 
 		const response = await this.sendReply(message, embed);
-		const func = () => response.delete().catch(() => undefined);
-		setTimeout(func, 5000);
+		if (response) {
+			const func = () => response.delete().catch(() => undefined);
+			setTimeout(func, 5000);
+		}
 	}
 
 	private images(messages: Message[]): Message[] {
