+++ src/commands/moderation/purge/purge-until.ts
@@ -60,8 +60,10 @@ export default class extends Command {
 			}
 
 			const response = await this.sendReply(message, embed);
-			const func = () => response.delete().catch(() => undefined);
-			setTimeout(func, 5000);
+			if (response) {
+				const func = () => response.delete().catch(() => undefined);
+				setTimeout(func, 5000);
+			}
 		}
 	}
 }
